update upvote/downvote system

+++ app/Http/Controllers/BlogController.php
@@ -56,31 +56,63 @@ class BlogController extends Controller
         $blog->views++;
         $blog->save();
         $comments = $blog->comments;
-        $upvotes = count(Votes::where('blog_id', $blog->id)->where('upvote', 1)->get());
-        $downvotes = count(Votes::where('blog_id', $blog->id)->where('downvote', 1)->get());
+        $upvotes = Votes::where('blog_id', $blog->id)->where('upvote', 1)->count();
+        $downvotes = Votes::where('blog_id', $blog->id)->where('downvote', 1)->count();
         return view('blogs.blog', compact('blog', 'image', 'video', 'audio', 'comments', 'upvotes', 'downvotes'));
     }
     
-    public function upvote(Blog $blog){
+    public function upvote(Request $request){
         $user = Auth::user();
-        $blog->votes()->detach($user->id);
-        $blog->votes()->attach([
-            $user->id => [
-                'upvote' => 1
-            ]
-        ]);
-        return redirect("/blog/".$blog->id);
+        $blog = Blog::find($request->blog_id);
+        if(!$blog) return abort(404);
+        $isUpvoted = $user->votes->filter(
+            function($i) use ($blog){
+                return $i->id == $blog->id && $i->pivot->upvote == 1;
+            })->count();
+        if($isUpvoted){
+            $blog->votes()->detach($user->id);
+            return 0;
+        }else{
+            $blog->votes()->detach($user->id);
+            $blog->votes()->attach([
+                $user->id => [
+                    'upvote' => 1
+                ]
+            ]);
+            return 1;
+        }
     }
-    public function downvote(Blog $blog){
+
+    public function getUpvotes(Blog $blog){
+        return Votes::where('blog_id', $blog->id)->where('upvote', 1)->count();
+    }
+
+    public function downvote(Request $request){
         $user = Auth::user();
-        $blog->votes()->detach($user->id);
-        $blog->votes()->attach([
-            $user->id => [
-                'downvote' => 1
-            ]
-        ]);
-        return redirect("/blog/".$blog->id);
+        $blog = Blog::find($request->blog_id);
+        if(!$blog) return abort(404);
+        $isDownvoted = $user->votes->filter(
+            function($i) use ($blog){
+                return $i->id == $blog->id && $i->pivot->downvote == 1;
+            })->count();
+        if($isDownvoted){
+            $blog->votes()->detach($user->id);
+            return 0;
+        }else{
+            $blog->votes()->detach($user->id);
+            $blog->votes()->attach([
+                $user->id => [
+                    'downvote' => 1
+                ]
+            ]);
+            return 1;
+        }
     }
+
+    public function getDownvotes(Blog $blog){
+        return Votes::where('blog_id', $blog->id)->where('downvote', 1)->count();
+    }
+
     public function comment(Blog $blog, Request $request){
         Comment::create([
             'user_id' => Auth::user()->id,
